backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
motor
pymongo
passlib
python-jose
langchain
langchain-community
langchain-huggingface
chromadb
sentence-transformers
pdfplumber
python-multipart
